requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
disnake==2.9.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
nodeenv==1.9.1
PSNAWP==2.1.0
pyrate-limiter==2.10.0
pyright==1.1.373
python-dotenv==1.0.1
requests==2.32.3
requests-ratelimiter==0.7.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
